requirements.txt
numpy==1.18.4
Pillow==7.1.2
PyOpenGL==3.1.5
